requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-docx==0.8.11
pandas==1.3.3
nltk==3.6.3
jieba==0.42.1
wordcloud==1.8.2.2
matplotlib==3.4.3
openpyxl==3.0.9
Pillow==8.4.0 